Rework selector children; always return string from *-text
Always returning a string from *-text revealed a bug which made
`elquery-children` not work on selectors, because elquery-children
expects a list of children instead of a single child.

Selectors used a single child previously, but now use a list of children,
just like nodes. That list is guaranteed to always contain one element.